Update README.md
 In the proposed system, the level 
of garbage in the dustbins is detected with the help of 
Sensor systems, and communicated to the authorized 
control room through GSM system. Microcontroller is 
used to interface the sensor system and GSM technology.
This will help to manage the garbage collection efficiently.
This application will be a all in one solution for garbage management as driver and garbage collector will be connected through each other via application portal.